Refactor: extracts `merge_ordered_arrays` to `physical-expr-common`
